create redaction map on every search, not log
