+++ astra/src/main/java/com/slack/astra/logstore/search/fieldRedaction/RedactedFieldReader.java
@@ -1,7 +1,10 @@
 package com.slack.astra.logstore.search.fieldRedaction;
 
+import com.slack.astra.metadata.fieldredaction.FieldRedactionMetadata;
 import com.slack.astra.metadata.fieldredaction.FieldRedactionMetadataStore;
 import java.io.IOException;
+import java.util.HashMap;
+
 import org.apache.lucene.codecs.StoredFieldsReader;
 import org.apache.lucene.index.StoredFieldVisitor;
 
@@ -13,17 +16,17 @@ import org.apache.lucene.index.StoredFieldVisitor;
 class RedactedFieldReader extends StoredFieldsReader {
 
   private final StoredFieldsReader in;
-  private final FieldRedactionMetadataStore fieldRedactionMetadataStore;
+  private final HashMap<String, FieldRedactionMetadata> fieldRedactionsMap;
 
   public RedactedFieldReader(
-      StoredFieldsReader in, FieldRedactionMetadataStore fieldRedactionMetadataStore) {
+      StoredFieldsReader in, HashMap<String, FieldRedactionMetadata> fieldRedactionsMap) {
     this.in = in;
-    this.fieldRedactionMetadataStore = fieldRedactionMetadataStore;
+    this.fieldRedactionsMap = fieldRedactionsMap;
   }
 
   @Override
   public StoredFieldsReader clone() {
-    return new RedactedFieldReader(in, fieldRedactionMetadataStore);
+    return new RedactedFieldReader(in, fieldRedactionsMap);
   }
 
   @Override
@@ -38,7 +41,7 @@ class RedactedFieldReader extends StoredFieldsReader {
 
   @Override
   public void document(int docID, StoredFieldVisitor visitor) throws IOException {
-    visitor = new RedactionStoredFieldVisitor(visitor, fieldRedactionMetadataStore);
+    visitor = new RedactionStoredFieldVisitor(visitor, fieldRedactionsMap);
     in.document(docID, visitor);
   }
 }
