+++ astra/src/main/java/com/slack/astra/logstore/search/fieldRedaction/RedactionStoredFieldVisitor.java
@@ -24,32 +24,9 @@ class RedactionStoredFieldVisitor extends StoredFieldVisitor {
   private final String redactedValue = "REDACTED";
 
   public RedactionStoredFieldVisitor(
-      final StoredFieldVisitor delegate, FieldRedactionMetadataStore fieldRedactionMetadataStore) {
+      final StoredFieldVisitor delegate, HashMap<String, FieldRedactionMetadata> fieldRedactionsMap) {
     super();
     this.delegate = delegate;
-
-    Map<String, FieldRedactionMetadata> fieldRedactionsMap = new HashMap<>();
-    fieldRedactionMetadataStore
-        .listSync()
-        .forEach(
-            redaction -> {
-              fieldRedactionsMap.put(redaction.getName(), redaction);
-            });
-
-    // TODO - do we need the listener at all if we listsync at the field level anyway?
-    AstraMetadataStoreChangeListener<FieldRedactionMetadata> listener =
-        new AstraMetadataStoreChangeListener() {
-          @Override
-          public void onMetadataStoreChanged(Object model) {
-            fieldRedactionMetadataStore
-                .listSync()
-                .forEach(
-                    redaction -> {
-                      fieldRedactionsMap.put(redaction.getName(), redaction);
-                    });
-          }
-        };
-    fieldRedactionMetadataStore.addListener(listener);
     this.fieldRedactionsMap = fieldRedactionsMap;
   }
