+++ astra/src/main/java/com/slack/astra/logstore/search/fieldRedaction/RedactionSubReaderWrapper.java
@@ -1,22 +1,47 @@
 package com.slack.astra.logstore.search.fieldRedaction;
 
+import com.slack.astra.metadata.core.AstraMetadataStoreChangeListener;
+import com.slack.astra.metadata.fieldredaction.FieldRedactionMetadata;
 import com.slack.astra.metadata.fieldredaction.FieldRedactionMetadataStore;
 import org.apache.lucene.index.FilterDirectoryReader;
 import org.apache.lucene.index.LeafReader;
+
+import java.util.HashMap;
 
 /**
  * RedactionSubReaderWrapper is called by the RedactionFilterDirectoryReader as a wrapper to the
  * reader, and creates a RedactionLeafReader.
  */
 class RedactionSubReaderWrapper extends FilterDirectoryReader.SubReaderWrapper {
-  private final FieldRedactionMetadataStore fieldRedactionMetadataStore;
+  private final HashMap<String, FieldRedactionMetadata> fieldRedactionsMap;
 
   public RedactionSubReaderWrapper(FieldRedactionMetadataStore fieldRedactionMetadataStore) {
-    this.fieldRedactionMetadataStore = fieldRedactionMetadataStore;
+    this.fieldRedactionsMap = new HashMap<>();
+    fieldRedactionMetadataStore
+            .listSync()
+            .forEach(
+                    redaction -> {
+                      fieldRedactionsMap.put(redaction.getName(), redaction);
+                    });
+
+    //TODO what is this listener for here?
+    AstraMetadataStoreChangeListener<FieldRedactionMetadata> listener =
+            new AstraMetadataStoreChangeListener() {
+              @Override
+              public void onMetadataStoreChanged(Object model) {
+                fieldRedactionMetadataStore
+                        .listSync()
+                        .forEach(
+                                redaction -> {
+                                  fieldRedactionsMap.put(redaction.getName(), redaction);
+                                });
+              }
+            };
+    fieldRedactionMetadataStore.addListener(listener);
   }
 
   @Override
   public LeafReader wrap(LeafReader reader) {
-    return new RedactionLeafReader(reader, this.fieldRedactionMetadataStore);
+    return new RedactionLeafReader(reader, this.fieldRedactionsMap);
   }
 }
